requirements.txt
# 基础依赖
python-dotenv>=1.0.0  # 环境变量管理
tqdm>=4.66.0         # 进度条显示
python-docx>=1.0.0   # Word文档处理

# OpenAI API
openai>=1.12.0       # OpenAI API客户端
httpx>=0.26.0        # HTTP客户端，用于OpenAI API请求

# Ollama API
ollama>=0.1.6        # Ollama客户端，使用最新版本 